+++ dashboards-observability/common/types/explorer.ts
@@ -187,11 +187,12 @@ export interface IConfigPanelOptions {
 export interface IConfigPanelOptionSection {
   name: string;
   component: null;
-  mapTo: 'mode';
+  mapTo: 'mode' | string;
   props?: any;
   isSingleSelection?: boolean;
   defaultState?: number;
   max?: number;
+  eleType?: string;
 }
 
 export interface IVisualizationTypeDefination {
