+++ dashboards-observability/public/components/event_analytics/explorer/visualizations/config_panel/config_panes/config_controls/index.ts
@@ -11,6 +11,5 @@ export { ConfigThresholds } from './config_thresholds';
 export { ConfigText } from './config_text';
 export { ConfigGaugeValueOptions } from './config_gauge_options';
 export { ColorPalettePicker } from './config_color_palette_picker';
-export { ConfigChartStyles } from './config_chart_styles';
+export { ConfigLineChartStyles } from './config_line_chart_styles';
 export { ConfigLegend } from './config_legend';
-export { ConfigGraphStyle } from './config_graph_style'
